+++ examples/d4rl/awac_pen/src/main.rs
@@ -10,11 +10,10 @@ use border_candle_agent::{
     Activation,
 };
 use border_core::{
-    generic_replay_buffer::{BatchBase, SimpleReplayBuffer},
-    record::Recorder,
-    Agent, Configurable, Env, Evaluator, ExperienceBufferBase, ReplayBufferBase, Trainer,
+    record::Recorder, Agent, Configurable, Env, Evaluator, ExperienceBuffer, ReplayBuffer, Trainer,
     TrainerConfig, TransitionBatch,
 };
+use border_generic_replay_buffer::{BatchBase, GenericReplayBuffer};
 use border_minari::{
     d4rl::pen::candle::{PenConverter, PenConverterConfig},
     MinariConverter, MinariDataset, MinariEnv, MinariEvaluator,
@@ -188,7 +187,7 @@ where
     E: Env + 'static,
     E::Obs: Into<Tensor>,
     E::Act: From<Tensor> + Into<Tensor>,
-    R: ReplayBufferBase + 'static,
+    R: ReplayBuffer + 'static,
     R::Batch: TransitionBatch,
     <R::Batch as TransitionBatch>::ObsBatch: Into<Tensor> + Clone,
     <R::Batch as TransitionBatch>::ActBatch: Into<Tensor> + Clone,
@@ -200,7 +199,7 @@ where
 fn create_replay_buffer<T>(
     converter: &mut T,
     dataset: &MinariDataset,
-) -> Result<SimpleReplayBuffer<T::ObsBatch, T::ActBatch>>
+) -> Result<GenericReplayBuffer<T::ObsBatch, T::ActBatch>>
 where
     T: MinariConverter,
     T::ObsBatch: BatchBase + Debug + Into<Tensor>,
@@ -215,7 +214,7 @@ where
 fn create_recorder<E, R>(config: &PenConfig) -> Result<Box<dyn Recorder<E, R>>>
 where
     E: Env + 'static,
-    R: ReplayBufferBase + 'static,
+    R: ReplayBuffer + 'static,
 {
     log::info!("Create recorder");
     if let Some(mlflow_run_name) = &config.args.mlflow_run_name {
@@ -282,7 +281,7 @@ where
     T::ObsBatch: std::fmt::Debug + Into<Tensor> + 'static + Clone,
     T::ActBatch: std::fmt::Debug + Into<Tensor> + 'static + Clone,
 {
-    let mut agent: Box<dyn Agent<MinariEnv<T>, SimpleReplayBuffer<T::ObsBatch, T::ActBatch>>> =
+    let mut agent: Box<dyn Agent<MinariEnv<T>, GenericReplayBuffer<T::ObsBatch, T::ActBatch>>> =
         create_agent(&config);
     let recorder = create_recorder(&config)?; // used for loading a trained model
     let mut evaluator = create_evaluator(&config.args, converter, &dataset, true)?;